backend/requirements.txt
fastapi
uvicorn==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
pydantic
python-dotenv==1.0.0
aiofiles==23.2.1
numpy
motor
pydantic[email]
mangum